% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritzrshiny.R
\docType{package}
\name{prioritizrshiny}
\alias{prioritizrshiny}
\alias{prioritizrshiny-package}
\title{prioritizrshiny: a package to use Shiy apps with  prioritizr}
\description{
The prioritizrshiny package provided Shiny apps
to work with prioritizr.
}
\section{prioritzrshiny functions}{

prshiny
}

