% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prshiny.R
\docType{methods}
\name{prshiny}
\alias{prshiny}
\alias{prioritizrshiny}
\title{Run one of the Shiy apps for prioritizr}
\arguments{
\item{x}{string for the app directory object.}

\item{...}{not used.}
}
\value{
Starts a Shiny app wrapper for prioritzr functions.
}
\description{
Description of the prshiny function.
}
\details{
Need to provide details for the function.
}
\examples{
# list available shiny apps
prshiny()

# run the basic app
prshiny("base_app")

}
\seealso{
\code{\link{problem}},  \code{\link{objectives}}, \code{\link{targets}}.
}
