% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prshiny_apps.R
\docType{methods}
\name{prshiny_apps}
\alias{prshiny_apps}
\title{Get a list of available Shiny apps}
\arguments{
\item{...}{not used.}
}
\value{
Returns a list of available Shiny apps.
}
\description{
Description of the prshiny function.
}
\details{
Need to provide details for the function.
}
\examples{
# list available shiny apps
prshiny_apps()

# run the basic app
# Not run
#prshiny("base_app")

}
\seealso{
\code{\link{prshiny}}.
}
