% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prshiny.R
\docType{methods}
\name{prshiny}
\alias{prshiny}
\alias{prioritizrshiny}
\title{Run one of the Shiy apps for prioritizrshiny}
\arguments{
\item{x}{string for the app directory object.}

\item{...}{not used.}
}
\value{
Starts a Shiny app wrapper for prioritzr functions.
}
\description{
Set targets expressed as the actual value of features in the study area 
that need to be represented in the prioritization. The argument to 
\code{x} is treated the same as for \code{\link{add_relative_targets}}.
}
\details{
Targets are used to specify the minimum amount or proportion of a feature's
distribution that needs to be protected. All conservation planning problems require 
adding targets with the exception of the maximum cover problem 
(see \code{\link{add_max_cover_objective}}), which maximizes all features 
in the solution and therefore does not require targets.
}
\examples{
# list available shiny apps
prshiny()

# run the basic app
prshiny("base_app")

}
\seealso{
\code{\link{problem}},  \code{\link{add_relative_targets}}, \code{\link{add_loglinear_targets}}.
}
